+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- use dynamodb's version of warehouse-models
 - [#5] Add collected docs and badges.
 
 [#5]: https://github.com/warehouseai/release-line/pull/5